requirements.txt
websockets==15.0.1
vosk==0.3.45
flask==3.1.2
google-generativeai==0.8.5
python-dotenv==1.2.1
